Med-Seo-Blog/pyproject.toml
[project]
name = "med-seo-blog"
version = "1.0.0"
description = "AI-powered Medical Tourism SEO & Blog Engine for ThaiTurk platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27.0",
    "google-genai>=1.0.0",
    "anthropic>=0.40.0",
]

[project.scripts]
serve = "uvicorn src.main:app --host 0.0.0.0 --port 8083 --reload"
